refactor: 변경된 경로 적용
