+++ starbucks/src/main/java/git_kkalnane/backend/starbucks/item/domain/ItemOption.java
@@ -1,6 +1,7 @@
 package git_kkalnane.backend.starbucks.item.domain;
 
 import git_kkalnane.backend.starbucks._global.entity.BaseTimeEntity;
+import git_kkalnane.backend.starbucks.item.domain.beverage.BeverageItemSyrup;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
