+++ starbucks/src/main/java/git_kkalnane/backend/starbucks/order/domain/OrderItem.java
@@ -1,8 +1,8 @@
 package git_kkalnane.backend.starbucks.order.domain;
 
 import git_kkalnane.backend.starbucks._global.entity.BaseTimeEntity;
-import git_kkalnane.backend.starbucks.item.domain.BeverageItem;
-import git_kkalnane.backend.starbucks.item.domain.DessertItem;
+import git_kkalnane.backend.starbucks.item.domain.beverage.BeverageItem;
+import git_kkalnane.backend.starbucks.item.domain.dessert.DessertItem;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
